Rename to LoadoutView and add some docs.
