+++ src/app/loadout/LoadoutView.m.scss
